+++ src/components/Header/Header.module.scss
@@ -1,47 +1,27 @@
-* {
-  box-sizing: border-box;
+@import "../../styles/assets.scss";
+
+.heading {
+  background-color: black;
+  text-align: center;
+  padding: 8px 0 4px 0;
 }
 
-.body {
-  margin: 0;
-  padding: 0;
-  font-family: "Hiragino Kaku Gothic Pro", "ヒラギノ角ゴ Pro W3", "Meiryo",
-    "ＭＳ Ｐゴシック", "Helvetica Neue", Helvetica, Arial, sans-serif;
-  background-color: #e6e6e6;
-}
-
-.header {
-  padding: 30px 4% 10px;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  background-color: #fff;
-  display: flex;
-  align-items: center;
-}
-
-.h1 {
-  margin: 0;
-  padding: 0;
-  font-size: 20px;
-}
-
-.a {
-  text-decoration: none;
-  color: #4b4b4b;
+.link {
+  color: white;
+  @include text-xl;
 }
 
 .nav {
-  margin: 0 0 0 auto;
+  background-color: $color-background;
 }
 
 .ul {
-  list-style: none;
-  margin: 0;
   display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
 .li {
-  margin: 0 0 0 15px;
-  font-size: 14px;
+  @include text-lg;
+  padding: 15px 33px;
 }
