+++ src/hooks/useSiteMetadata.jsx
@@ -5,36 +5,34 @@ const useSiteMetadata = () => {
   return site.siteMetadata
 }
 
-
 // とりあえず必要そうなサイトデータだけ取る
-const siteMetadataQuery =
-    graphql`
-      query SiteMetaDataQuery {
-        site {
-          siteMetadata {
-            title
-            siteUrl
-            description
-            author {
-              name
-              summary
-            }
-            social {
-              twitter
-              github
-              instagram
-            }
-            headerMenu {
-              name
-              slug
-            }
-            footerMenu {
-              name
-              slug
-            }
-          }
+const siteMetadataQuery = graphql`
+  query SiteMetaDataQuery {
+    site {
+      siteMetadata {
+        title
+        siteUrl
+        description
+        author {
+          name
+          summary
+        }
+        social {
+          twitter
+          github
+          instagram
+        }
+        headerMenu {
+          name
+          slug
+        }
+        footerMenu {
+          name
+          slug
         }
       }
-    `
+    }
+  }
+`
 
 export default useSiteMetadata
