+++ src/styles/assets.scss
@@ -1,0 +1,3 @@
+@import "reset";
+@import "variables";
+@import "mixins";
