+++ src/styles/global.scss
@@ -1,0 +1,4 @@
+// global css
+body {
+  font-family: "Volte", sans-serif;
+}
