+++ src/styles/variables.scss
@@ -1,0 +1,8 @@
+/*
+  scss variables
+*/
+
+// color
+$color-font: #000;
+$color-accent: #FF9D43;
+$color-background: #fff;
